+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 with open("README.md") as f:
     long_description = f.read()
@@ -12,7 +12,8 @@ setup(
     long_description=long_description,
     author="Austin Long",
     author_email="austin@austinlong.dev",
-    packages=["zderad"],  # same as name
+    packages=find_packages(),
+    install_requires=["colored"],
     entry_points={
         "console_scripts": ["zderad=zderad.main:main"],
     },
