+++ zderad/directives/__init__.py
@@ -1,0 +1,4 @@
+from .incfile import IncludeFileDirective
+from .incimgs import IncludeOutputImagesDirective
+
+__all__ = [IncludeFileDirective, IncludeOutputImagesDirective]
